Split new major types into their own files (Aziz policy)
PhoneForwardingMode, MacPresenceDecisionCache, and ScreenLockObserver
each move to a dedicated file; MacPresenceMonitor.swift keeps only the
monitor and its live-provider extension. No behavior change.
