+++ Sources/MacPresenceMonitor.swift
@@ -92,85 +92,6 @@ struct MacPresenceMonitor {
     }
 }
 
-/// Bounds live presence sampling under notification bursts while the user is
-/// AT the Mac (the suppression path, which deliberately never consumes a
-/// send-throttle slot in `PhonePushClient.forward`).
-///
-/// Invariant, chosen explicitly: any bounded-work design serves a decision up
-/// to ``ttl`` stale on one side of an active/away transition. The staleness
-/// goes on the SUPPRESSION side only:
-///
-/// - ACTIVE decisions are reused for up to ``ttl``. A burst while the user is
-///   at the Mac samples WindowServer/AppKit/HID at most once per TTL instead
-///   of per notification. This is the typing-adjacent main-actor hot path the
-///   send throttle cannot protect, because suppressed notifications must not
-///   consume send slots.
-/// - AWAY decisions are NEVER reused. A forwarding decision is always fresh,
-///   so a stale away answer can never leak a push to the phone after the
-///   user has returned to the Mac.
-///
-/// Cost of the chosen staleness: a notification arriving within ``ttl`` after
-/// the user locks/leaves can be suppressed once. The user was provably at the
-/// Mac within the last second (hardware input, unlocked screen) and saw the
-/// notification there; suppressed pushes are never retroactively sent by
-/// design, and the next notification more than ``ttl`` later re-samples
-/// fresh.
-struct MacPresenceDecisionCache {
-    static let ttl: TimeInterval = 1.0
-
-    private var last: MacPresenceMonitor.Decision?
-
-    mutating func decision(from monitor: MacPresenceMonitor) -> MacPresenceMonitor.Decision {
-        let now = monitor.now()
-        if let last,
-           last.isActive,
-           now >= last.evaluatedAt,
-           now.timeIntervalSince(last.evaluatedAt) < Self.ttl {
-            return last
-        }
-        let fresh = monitor.evaluate()
-        last = fresh
-        return fresh
-    }
-}
-
-/// There is no public synchronous "is the screen locked" query on macOS. Two
-/// de-facto sources exist: the `CGSSessionScreenIsLocked` key in
-/// `CGSessionCopyCurrentDictionary()` and the `com.apple.screenIsLocked` /
-/// `com.apple.screenIsUnlocked` distributed notifications. Either can be
-/// absent in a given macOS version or session context, so the live provider
-/// ORs both (`MacPresenceMonitor.consoleSessionActiveAndUnlocked`). If both
-/// miss a lock, the failure mode is bounded: the 120 s hardware-idle rule
-/// flips the Mac to away on its own shortly after the user leaves.
-@MainActor
-final class ScreenLockObserver {
-    static let shared = ScreenLockObserver()
-
-    private(set) var isLockedObserved = false
-
-    /// Retained for the life of the process; the observer is a singleton
-    /// whose lifetime is the app's.
-    private var observerTokens: [any NSObjectProtocol] = []
-
-    private init() {
-        let center = DistributedNotificationCenter.default()
-        observerTokens.append(center.addObserver(
-            forName: Notification.Name("com.apple.screenIsLocked"),
-            object: nil,
-            queue: .main
-        ) { [weak self] _ in
-            MainActor.assumeIsolated { self?.isLockedObserved = true }
-        })
-        observerTokens.append(center.addObserver(
-            forName: Notification.Name("com.apple.screenIsUnlocked"),
-            object: nil,
-            queue: .main
-        ) { [weak self] _ in
-            MainActor.assumeIsolated { self?.isLockedObserved = false }
-        })
-    }
-}
-
 extension MacPresenceMonitor {
     /// Production monitor backed by the real WindowServer/HID signals. The
     /// live monitor is owned and evaluated by `PhonePushClient` on the main
